+++ src/tmui/components/tm-picker/tm-picker.vue
@@ -8,7 +8,7 @@
     <!-- #ifndef APP-NVUE -->
     <slot></slot>
     <!-- #endif -->
-    <tm-drawer :inContent="props.inContent" :disabbleScroll="true" :round="props.round" ref="drawer" :height="dHeight"
+    <tm-drawer :zIndex="props.zIndex" :inContent="props.inContent" :disabbleScroll="true" :round="props.round" ref="drawer" :height="dHeight"
       :closeable="true" :overlayClick="aniover" @open="drawerOpen" @cancel="cancel" @ok="confirm" :show="showCity"
       @close="closeDrawer"
       title="请选择" ok-text="确认">
@@ -165,7 +165,11 @@ const props = defineProps({
   disabled:{
     type:Boolean,
     default:false
-  }
+  },
+  zIndex: {
+    type: [Number, String],
+    default: 999,
+  },
 });
 const showCity = ref(true);
 const _colIndex: Ref<Array<number>> = ref([]);