+++ src/tmui/tool/theme/theme.ts
@@ -10,16 +10,7 @@ import { colortool } from './colortool';
 import  { cssStyleConfig, cssstyle, colorThemeType, cssDirection, linearDirection, linearDeep, linearDirectionType } from '../lib/interface';
 //导入用户自定义的主题色值。
 // import { theme } from '../../../theme/index';
-
-let theme = uni.$tm?.config?.theme
-// #ifdef APP
-if(!uni.$tm?.config?.theme){
-	theme = uni.getStorageSync('$tm')||{}
-}
-// #endif
-// #ifndef APP
-theme = theme||{}
-// #endif
+let theme = (uni?.$tm?.config?.theme??uni.getStorageSync('$tm'))||{}
 
 var colors: Array<colorThemeType> = [];
 var colorObj: any = {